Fix for bottom toolbar staying in middle of page after entering text.
When entering text into a text area on Android the bottom toolbar will re-position
itself so that the user can see the text area while they enter text with the
virtual keyboard.
If the bottom toolbar, which appears above the virtual keyboard, hides the
text area then the bottom toolbar will be moved up so that the user can see
the text they are entering into the text area.

After the virtual keyboard is dismissed the bottom toolbar is positioned incorrectly.
It's not at the bottom of the page and it's buttons are not responsive.