Fix for incorrect labels in Galileo.
[#52345779]

Zeus hand-edited gravity.js, adding 'activeDatasets' arrays to each pane specification and (incorrectly) explicitly labeling some graphs as "Position" graphs.

This commit just restores gravity.js to a compiled version of gravity.coffee.

In order for the tagging tool to work correctly, it seems the datadef must now have its 'isActive' flag set to YES. The previous commit causes the tagging tool to automatically mark its datadef as active, so that we don't need the 'activeDatadefs' that Zeus added to gravity.js. The commit before that sets the tableController to automatically use the xLabel/yLabel of a datadef if there is no explicit override in the table pane specification.