Tagging tool should set isActive on its datadef.
[#52345779]

Without this behavior, clicks on datadefs are ignored and the tagging tool doesn't work correctly. The previous solution was to explicitly add the datadef to the 'activeDatasets' key in the JSON specification of each pane that used the tagging tool, or else hope that there is only one datadef so that it's active by default.