PointView: don't re-test hover on mouseup/touchend
[#61991698]

This leads to a persistent, undismissable hover state when touch events are used (because the code sets isHovered on touchend as well as mouseup--but when using touch events you'll never see the unhovering mouseout!). Could scope this logic only to mouseup, but on second thought, it's actually better UX to require explicit hover to get the hover tooltip back after you've successfully placed the point where you want.