Add hack to solve touchEnd event not firing during line-construction
This is an ugly hack to solve a bug in the Android app, where the touchEnd
event is not triggered during a construct-a-line sequence. This bug occurs
in the Chromium-based WebView used in Android 4.4 KitKat. It seems to be
related to https://code.google.com/p/chromium/issues/detail?id=152913 and
https://code.google.com/p/android/issues/detail?id=19827 but is not exactly
the same, and does not seem to be resolved by adding touch.preventDefault()
to touchStart.

This hack simply starts a timer on touchStart and, if no movement occurs for
750 ms, automatically calls touchEnd.

It also uses a touchStarted flag to prevent accidentally calling touchEnd
twice for the same touch.