Make it possible to show a 'Home' button on the toolbar
This feature is for the Algebra mobile app, which will contain
a menu page from which a user can access activities. The
button will allow the user to return to the menu page from
within the app.

This is controlled by two new properties, window.showHomeButton
(default false) and window.activityHome (default '/index.html').

This matches the pattern set by showEditButton etc, which allows
these settings to be in the html page.

The adjustLayout bindings to move the other two toolbar buttons
were very annoying... Without the init() function, the behavior
was non-deterministic.