Remove broken touchBoundaryFrame fix
This fix was added pre SC 1.9, and while it made the button clickable,
it also removed the touchBoundary slop that had been built into the
view.

A better fix for the bug was introduced in SC 1.9

http://blog.sproutcore.com/announcing-sproutcore-1-9/

This fix keeps the touchBoundary slop, which allows a user to slightly
drag their finger off the button without causing a touchCancel.

(Note, this touch boundary does not actually affect the initial hit area.)