Widen touchBoundary on Submit and Next buttons
This touchBoundary allows the user's finger to slip off the button
slightly without causing a touchCancel event. Note that it does not
affect the initial hit target.

The touchBoundary now works again on the Submit button after ea8209b91

This widens the touchBoundary from the default of 25px to 50px.